Minutes calculation in Timer fixed. COmplexity for all levels changed. Order of levels hardcoded.

+++ CodingDojo/games/kata/src/main/java/com/codenjoy/dojo/kata/model/Timer.java
@@ -29,7 +29,7 @@ import java.util.Calendar;
  * Created by indigo on 2017-04-03.
  */
 public class Timer {
-    public static int ONE_MINUTE_IN_MILLS = 60000;
+    public static double ONE_MINUTE_IN_MILLS = 60000;
 
     private long start = now();
